+++ .storybook/main.js
@@ -1,30 +1,28 @@
-const path = require('path');
+const path = require("path");
 
 module.exports = {
-  stories: ['../src/**/*.stories.(js|mdx)', './welcome-page/welcome.stories.js', '../docs/*.stories.mdx'],
+  stories: [
+    "../src/**/*.stories.(js|mdx)",
+    "./welcome-page/welcome.stories.js",
+    "../docs/*.stories.mdx",
+  ],
   addons: [
-    '@storybook/addon-actions/register',
-    '@storybook/addon-knobs/register',
-    '@storybook/addon-notes/register',
-    '@storybook/addon-viewport/register',
-    '@storybook/addon-a11y/register',
-    '@storybook/addon-docs',
-    '@storybook/addon-google-analytics/register',
-    '@storybook/addon-links',
-    './.storybook/theme-selector/register',
+    "@storybook/addon-actions/register",
+    "@storybook/addon-knobs/register",
+    "@storybook/addon-notes/register",
+    "@storybook/addon-viewport/register",
+    "@storybook/addon-a11y/register",
+    "@storybook/addon-docs",
+    "@storybook/addon-google-analytics/register",
+    "@storybook/addon-links",
+    "./.storybook/theme-selector/register",
   ],
   webpackFinal: async (config, { configType }) => {
-    config.module.rules.push(
-      {
-        test: /\.scss$/, use: ['style-loader', 'css-loader', 'sass-loader']
-      }
-    );
-  
     config.resolve = {
       alias: {
-        helpers: path.resolve(__dirname, '__helpers__/')
+        helpers: path.resolve(__dirname, "__helpers__/"),
       },
-      extensions: ['.js']
+      extensions: [".js"],
     };
     return config;
   },