+++ .storybook/with-global-styles.js
@@ -1,0 +1,11 @@
+import React from "react";
+import GlobalStyle from "../src/style/global-style";
+
+export default function withGlobalStyles(content) {
+  return (
+    <>
+      <GlobalStyle />
+      {content()}
+    </>
+  );
+}